DPDV-2608: Introduce alias s1query
