+++ e2e/global.setup.ts
@@ -73,7 +73,7 @@ setup('login and create account', async ({ page }) => {
     const { DATASET_URL: datasetUrl, DATASET_LOG_ACCESS_READ: datasetReadKey, DATASET_LOG_ACCESS_WRITE: datasetWriteKey } = process.env;
 
     // Fill in values
-    const accountName = 'QaTr' + (Math.random() * 1e18)
+    const accountName = 'E2ET' + (Math.random() * 1e18)
     console.log("Create account: ", accountName);
     await locAccount.fill(accountName);
     await locUrl.fill(datasetUrl || '');
